ST6RI-363 Moved KerMLValidator.multiplicityRange to FeatureUtil.
- Also moved addMultiplicityTo from TypeUtil to FeatureUtil.
